cut operator
Cut operator is useful when there is confidence other alternatives
are not worth to check: it marks whole `alt` as failed if current
alternative will fail.

It also could simplify grammar, as in test_logo.py where `name` rule
is implemented now without keyword check.

It is well known and wide-spread in PEG implementations:
https://peps.python.org/pep-0617/#e-5
https://github.com/yhirose/cpp-peglib (Cut operator)
https://ceur-ws.org/Vol-1269/paper232.pdf
https://bookish.press/hcpl/chapter6 (Cut commits to the current alternative...)